src/frontend/requirements.txt
streamlit==1.12.0
requests==2.26.0
pandas==1.3.3
plotly==5.3.1
python-dotenv==0.19.0
Pillow==8.3.2
beautifulsoup4==4.10.0
markdown==3.3.4
python-docx==0.8.11
openpyxl==3.0.9
PyPDF2==1.26.0
wordcloud==1.8.1
matplotlib==3.4.3
toml==0.10.2
sqlalchemy==1.4.25